Initial Commit.
Figuring out hinotify interface. I think the sane way to get this
working is to watch the directory where the files will be created, and
then proceed to add each created new file (along with prexisiting
files) to a list of watched files. Each file will have it's own
callback.

I don't know if I need to have each thing to it's own shared queue or
not so that I don't have wonky mixed lines.