Allowing IO actions in handler, updated types.
An alternative approach is to try to isolate the state transforms from
IO actions, and then always execute the IO actions after updating the
state first. However, the appropriate state updates rely on IO actions
(like the length of the file). Not clear if that's a good reason to
return IO State, or if it's actually call for a loop "Update
State, if yielded IO action, execute, go back, stop if no change"